+++ lib/recite_csv/reader/row.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module ReciteCSV
+  module Reader
+    module Row
+    end
+  end
+end
+
+require "recite_csv/reader/row/base"