+++ lib/recite_csv/row.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module ReciteCSV
-  module Row
-  end
-end
-
-require "recite_csv/row/base"