+++ spec/recite_csv/reader/core_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe ReciteCSV::Reader::Core do
   let(:dummy_class) do
     Class.new do
       include ReciteCSV::Reader::Core
-      const_set(:Row, Class.new(ReciteCSV::Row::Base))
+      const_set(:Row, Class.new(ReciteCSV::Reader::Row::Base))
       const_set(:DEFAULT_CSV_OPTIONS, headers: :first_row)
     end
   end
@@ -28,7 +28,7 @@ RSpec.describe ReciteCSV::Reader::Core do
     context "call with block" do
       it "enumerate with row object" do
         reader.each do |row|
-          expect(row).to be_a ReciteCSV::Row::Base
+          expect(row).to be_a ReciteCSV::Reader::Row::Base
           expect(row["COL1"]).to eq "V1"
           expect(row["COL2"]).to eq "V2"
         end